feat: Customising ConfigurationBinder to handle immutablelist
