+++ src/Uno.Extensions.Configuration/ConfigurationBinder.cs
@@ -1,18 +1,10 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
+﻿
 
-using System;
-using System.Collections.Generic;
+using System.Collections.Immutable;
 using System.ComponentModel;
-using System.Diagnostics.CodeAnalysis;
-using System.Linq;
-using System.Reflection;
 
-namespace Microsoft.Extensions.Configuration;
+namespace Uno.Extensions.Configuration.Internal;
 
-/// <summary>
-/// Static helper class that allows binding strongly typed objects to configuration values.
-/// </summary>
 public static class ConfigurationBinder
 {
 	private const BindingFlags DeclaredOnlyLookup = BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance | BindingFlags.Static | BindingFlags.DeclaredOnly;
@@ -28,8 +20,7 @@ public static class ConfigurationBinder
 	/// <typeparam name="T">The type of the new instance to bind.</typeparam>
 	/// <param name="configuration">The configuration instance to bind.</param>
 	/// <returns>The new instance of T if successful, default(T) otherwise.</returns>
-	[RequiresUnreferencedCode(TrimmingWarningMessage)]
-	public static T Get<[DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.All)] T>(this IConfiguration configuration)
+	public static T Get<T>(this IConfiguration configuration)
 		=> configuration.Get<T>(_ => { });
 
 	/// <summary>
@@ -41,8 +32,8 @@ public static class ConfigurationBinder
 	/// <param name="configuration">The configuration instance to bind.</param>
 	/// <param name="configureOptions">Configures the binder options.</param>
 	/// <returns>The new instance of T if successful, default(T) otherwise.</returns>
-	[RequiresUnreferencedCode(TrimmingWarningMessage)]
-	public static T Get<[DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.All)] T>(this IConfiguration configuration, Action<BinderOptions> configureOptions)
+
+	public static T Get<T>(this IConfiguration configuration, Action<BinderOptions> configureOptions)
 	{
 		if (configuration == null)
 		{
@@ -52,7 +43,7 @@ public static class ConfigurationBinder
 		object result = configuration.Get(typeof(T), configureOptions);
 		if (result == null)
 		{
-			return default(T);
+			return default(T)!;
 		}
 		return (T)result;
 	}
@@ -65,7 +56,7 @@ public static class ConfigurationBinder
 	/// <param name="configuration">The configuration instance to bind.</param>
 	/// <param name="type">The type of the new instance to bind.</param>
 	/// <returns>The new instance if successful, null otherwise.</returns>
-	[RequiresUnreferencedCode(TrimmingWarningMessage)]
+
 	public static object Get(this IConfiguration configuration, Type type)
 		=> configuration.Get(type, _ => { });
 
@@ -78,11 +69,11 @@ public static class ConfigurationBinder
 	/// <param name="type">The type of the new instance to bind.</param>
 	/// <param name="configureOptions">Configures the binder options.</param>
 	/// <returns>The new instance if successful, null otherwise.</returns>
-	[RequiresUnreferencedCode(TrimmingWarningMessage)]
+
 	public static object Get(
 		this IConfiguration configuration,
-		[DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.All)]
-		Type type,
+
+			Type type,
 		Action<BinderOptions> configureOptions)
 	{
 		if (configuration == null)
@@ -92,7 +83,7 @@ public static class ConfigurationBinder
 
 		var options = new BinderOptions();
 		configureOptions?.Invoke(options);
-		return BindInstance(type, instance: null, config: configuration, options: options);
+		return BindInstance(type, instance: null!, config: configuration, options: options);
 	}
 
 	/// <summary>
@@ -101,7 +92,7 @@ public static class ConfigurationBinder
 	/// <param name="configuration">The configuration instance to bind.</param>
 	/// <param name="key">The key of the configuration section to bind.</param>
 	/// <param name="instance">The object to bind.</param>
-	[RequiresUnreferencedCode(InstanceGetTypeTrimmingWarningMessage)]
+
 	public static void Bind(this IConfiguration configuration, string key, object instance)
 		=> configuration.GetSection(key).Bind(instance);
 
@@ -110,7 +101,7 @@ public static class ConfigurationBinder
 	/// </summary>
 	/// <param name="configuration">The configuration instance to bind.</param>
 	/// <param name="instance">The object to bind.</param>
-	[RequiresUnreferencedCode(InstanceGetTypeTrimmingWarningMessage)]
+
 	public static void Bind(this IConfiguration configuration, object instance)
 		=> configuration.Bind(instance, o => { });
 
@@ -120,7 +111,7 @@ public static class ConfigurationBinder
 	/// <param name="configuration">The configuration instance to bind.</param>
 	/// <param name="instance">The object to bind.</param>
 	/// <param name="configureOptions">Configures the binder options.</param>
-	[RequiresUnreferencedCode(InstanceGetTypeTrimmingWarningMessage)]
+
 	public static void Bind(this IConfiguration configuration, object instance, Action<BinderOptions> configureOptions)
 	{
 		if (configuration == null)
@@ -143,10 +134,10 @@ public static class ConfigurationBinder
 	/// <param name="configuration">The configuration.</param>
 	/// <param name="key">The key of the configuration section's value to convert.</param>
 	/// <returns>The converted value.</returns>
-	[RequiresUnreferencedCode(TrimmingWarningMessage)]
-	public static T GetValue<[DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.All)] T>(this IConfiguration configuration, string key)
+
+	public static T GetValue<T>(this IConfiguration configuration, string key)
 	{
-		return GetValue(configuration, key, default(T));
+		return GetValue(configuration, key, default(T))!;
 	}
 
 	/// <summary>
@@ -157,10 +148,10 @@ public static class ConfigurationBinder
 	/// <param name="key">The key of the configuration section's value to convert.</param>
 	/// <param name="defaultValue">The default value to use if no value is found.</param>
 	/// <returns>The converted value.</returns>
-	[RequiresUnreferencedCode(TrimmingWarningMessage)]
-	public static T GetValue<[DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.All)] T>(this IConfiguration configuration, string key, T defaultValue)
+
+	public static T GetValue<T>(this IConfiguration configuration, string key, T defaultValue)
 	{
-		return (T)GetValue(configuration, typeof(T), key, defaultValue);
+		return (T)GetValue(configuration, typeof(T), key, defaultValue!);
 	}
 
 	/// <summary>
@@ -170,14 +161,14 @@ public static class ConfigurationBinder
 	/// <param name="type">The type to convert the value to.</param>
 	/// <param name="key">The key of the configuration section's value to convert.</param>
 	/// <returns>The converted value.</returns>
-	[RequiresUnreferencedCode(TrimmingWarningMessage)]
+
 	public static object GetValue(
 		this IConfiguration configuration,
-		[DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.All)]
-		Type type,
+
+			Type type,
 		string key)
 	{
-		return GetValue(configuration, type, key, defaultValue: null);
+		return GetValue(configuration, type, key, defaultValue: null!);
 	}
 
 	/// <summary>
@@ -188,11 +179,11 @@ public static class ConfigurationBinder
 	/// <param name="key">The key of the configuration section's value to convert.</param>
 	/// <param name="defaultValue">The default value to use if no value is found.</param>
 	/// <returns>The converted value.</returns>
-	[RequiresUnreferencedCode(TrimmingWarningMessage)]
+
 	public static object GetValue(
 		this IConfiguration configuration,
-		[DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.All)]
-		Type type, string key,
+
+			Type type, string key,
 		object defaultValue)
 	{
 		IConfigurationSection section = configuration.GetSection(key);
@@ -204,7 +195,7 @@ public static class ConfigurationBinder
 		return defaultValue;
 	}
 
-	[RequiresUnreferencedCode(PropertyTrimmingWarningMessage)]
+
 	private static void BindNonScalar(this IConfiguration configuration, object instance, BinderOptions options)
 	{
 		if (instance != null)
@@ -224,9 +215,7 @@ public static class ConfigurationBinder
 
 				if (missingPropertyNames.Count > 0)
 				{
-					throw new InvalidOperationException(SR.Format(SR.Error_MissingConfig,
-						nameof(options.ErrorOnUnknownConfiguration), nameof(BinderOptions), instance.GetType(),
-						string.Join(", ", missingPropertyNames)));
+					throw new InvalidOperationException("error");
 				}
 			}
 
@@ -237,7 +226,7 @@ public static class ConfigurationBinder
 		}
 	}
 
-	[RequiresUnreferencedCode(PropertyTrimmingWarningMessage)]
+
 	private static void BindProperty(PropertyInfo property, object instance, IConfiguration config, BinderOptions options)
 	{
 		// We don't support set only, non public, or indexer properties
@@ -266,7 +255,6 @@ public static class ConfigurationBinder
 		}
 	}
 
-	[RequiresUnreferencedCode("Cannot statically analyze what the element type is of the object collection in type so its members may be trimmed.")]
 	private static object BindToCollection(Type type, IConfiguration config, BinderOptions options)
 	{
 		Type genericType = typeof(List<>).MakeGenericType(type.GenericTypeArguments[0]);
@@ -276,15 +264,14 @@ public static class ConfigurationBinder
 	}
 
 	// Try to create an array/dictionary instance to back various collection interfaces
-	[RequiresUnreferencedCode("In case type is a Dictionary, cannot statically analyze what the element type is of the value objects in the dictionary so its members may be trimmed.")]
 	private static object AttemptBindToCollectionInterfaces(
-		[DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.All)]
-		Type type,
+
+			Type type,
 		IConfiguration config, BinderOptions options)
 	{
 		if (!type.IsInterface)
 		{
-			return null;
+			return null!;
 		}
 
 		Type collectionInterface = FindOpenGenericInterface(typeof(IReadOnlyList<>), type);
@@ -332,13 +319,13 @@ public static class ConfigurationBinder
 			return BindToCollection(type, config, options);
 		}
 
-		return null;
+		return null!;
 	}
 
-	[RequiresUnreferencedCode(TrimmingWarningMessage)]
+
 	private static object BindInstance(
-		[DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.All)]
-		Type type,
+
+			Type type,
 		object instance, IConfiguration config, BinderOptions options)
 	{
 		// if binding IConfigurationSection, break early
@@ -348,10 +335,10 @@ public static class ConfigurationBinder
 		}
 
 		var section = config as IConfigurationSection;
-		string configValue = section?.Value;
+		string configValue = section?.Value!;
 		object convertedValue;
 		Exception error;
-		if (configValue != null && TryConvertValue(type, configValue, section.Path, out convertedValue, out error))
+		if (configValue != null && TryConvertValue(type, configValue, section?.Path!, out convertedValue, out error))
 		{
 			if (error != null)
 			{
@@ -389,16 +376,24 @@ public static class ConfigurationBinder
 			}
 			else
 			{
-				// See if its an ICollection
-				collectionInterface = FindOpenGenericInterface(typeof(ICollection<>), type);
-				if (collectionInterface != null)
+				var immutableInterface = FindOpenGenericInterface(typeof(IImmutableList<>), type);
+				if (immutableInterface != null)
 				{
-					BindCollection(instance, collectionInterface, config, options);
+					instance = BindImmutableList(instance, immutableInterface, config, options);
 				}
-				// Something else
 				else
 				{
-					BindNonScalar(config, instance, options);
+					// See if its an ICollection
+					collectionInterface = FindOpenGenericInterface(typeof(ICollection<>), type);
+					if (collectionInterface != null)
+					{
+						BindCollection(instance, collectionInterface, config, options);
+					}
+					// Something else
+					else
+					{
+						BindNonScalar(config, instance, options);
+					}
 				}
 			}
 		}
@@ -406,18 +401,20 @@ public static class ConfigurationBinder
 		return instance;
 	}
 
-	private static object CreateInstance([DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.PublicConstructors | DynamicallyAccessedMemberTypes.NonPublicConstructors)] Type type)
+	private static object CreateInstance(
+
+		Type type)
 	{
 		if (type.IsInterface || type.IsAbstract)
 		{
-			throw new InvalidOperationException(SR.Format(SR.Error_CannotActivateAbstractOrInterface, type));
+			throw new InvalidOperationException("error");
 		}
 
 		if (type.IsArray)
 		{
 			if (type.GetArrayRank() > 1)
 			{
-				throw new InvalidOperationException(SR.Format(SR.Error_UnsupportedMultidimensionalArray, type));
+				throw new InvalidOperationException("error");
 			}
 
 			return Array.CreateInstance(type.GetElementType(), 0);
@@ -428,7 +425,7 @@ public static class ConfigurationBinder
 			bool hasDefaultConstructor = type.GetConstructors(DeclaredOnlyLookup).Any(ctor => ctor.IsPublic && ctor.GetParameters().Length == 0);
 			if (!hasDefaultConstructor)
 			{
-				throw new InvalidOperationException(SR.Format(SR.Error_MissingParameterlessConstructor, type));
+				throw new InvalidOperationException("error");
 			}
 		}
 
@@ -438,15 +435,13 @@ public static class ConfigurationBinder
 		}
 		catch (Exception ex)
 		{
-			throw new InvalidOperationException(SR.Format(SR.Error_FailedToActivate, type), ex);
+			throw new InvalidOperationException("error", ex);
 		}
 	}
 
-	[RequiresUnreferencedCode("Cannot statically analyze what the element type is of the value objects in the dictionary so its members may be trimmed.")]
 	private static void BindDictionary(
 		object dictionary,
-		[DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.PublicProperties | DynamicallyAccessedMemberTypes.NonPublicProperties)]
-		Type dictionaryType,
+			Type dictionaryType,
 		IConfiguration config, BinderOptions options)
 	{
 		// IDictionary<K,V> is guaranteed to have exactly two parameters
@@ -465,7 +460,7 @@ public static class ConfigurationBinder
 		{
 			object item = BindInstance(
 				type: valueType,
-				instance: null,
+				instance: null!,
 				config: child,
 				options: options);
 			if (item != null)
@@ -484,11 +479,9 @@ public static class ConfigurationBinder
 		}
 	}
 
-	[RequiresUnreferencedCode("Cannot statically analyze what the element type is of the object collection so its members may be trimmed.")]
 	private static void BindCollection(
 		object collection,
-		[DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.PublicMethods | DynamicallyAccessedMemberTypes.NonPublicMethods)]
-		Type collectionType,
+			Type collectionType,
 		IConfiguration config, BinderOptions options)
 	{
 		// ICollection<T> is guaranteed to have exactly one parameter
@@ -501,7 +494,7 @@ public static class ConfigurationBinder
 			{
 				object item = BindInstance(
 					type: itemType,
-					instance: null,
+					instance: null!,
 					config: section,
 					options: options);
 				if (item != null)
@@ -515,7 +508,38 @@ public static class ConfigurationBinder
 		}
 	}
 
-	[RequiresUnreferencedCode("Cannot statically analyze what the element type is of the Array so its members may be trimmed.")]
+	private static object BindImmutableList(
+		object collection,
+			Type collectionType,
+		IConfiguration config, BinderOptions options)
+	{
+		// ICollection<T> is guaranteed to have exactly one parameter
+		Type itemType = collectionType.GenericTypeArguments[0];
+		MethodInfo addMethod = collectionType.GetMethod("Add", DeclaredOnlyLookup);
+
+		foreach (IConfigurationSection section in config.GetChildren())
+		{
+			try
+			{
+				object item = BindInstance(
+					type: itemType,
+					instance: null!,
+					config: section,
+					options: options);
+				if (item != null)
+				{
+					collection = addMethod.Invoke(collection, new[] { item });
+				}
+			}
+			catch
+			{
+			}
+		}
+
+		return collection;
+	}
+
+
 	private static Array BindArray(Array source, IConfiguration config, BinderOptions options)
 	{
 		IConfigurationSection[] children = config.GetChildren().ToArray();
@@ -535,7 +559,7 @@ public static class ConfigurationBinder
 			{
 				object item = BindInstance(
 					type: elementType,
-					instance: null,
+					instance: null!,
 					config: children[i],
 					options: options);
 				if (item != null)
@@ -551,14 +575,14 @@ public static class ConfigurationBinder
 		return newArray;
 	}
 
-	[RequiresUnreferencedCode(TrimmingWarningMessage)]
+
 	private static bool TryConvertValue(
-		[DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.All)]
-		Type type,
+
+			Type type,
 		string value, string path, out object result, out Exception error)
 	{
-		error = null;
-		result = null;
+		error = null!;
+		result = null!;
 		if (type == typeof(object))
 		{
 			result = value;
@@ -583,7 +607,7 @@ public static class ConfigurationBinder
 			}
 			catch (Exception ex)
 			{
-				error = new InvalidOperationException(SR.Format(SR.Error_FailedBinding, path, type), ex);
+				error = new InvalidOperationException("error", ex);
 			}
 			return true;
 		}
@@ -596,7 +620,7 @@ public static class ConfigurationBinder
 			}
 			catch (FormatException ex)
 			{
-				error = new InvalidOperationException(SR.Format(SR.Error_FailedBinding, path, type), ex);
+				error = new InvalidOperationException("error", ex);
 			}
 			return true;
 		}
@@ -604,10 +628,10 @@ public static class ConfigurationBinder
 		return false;
 	}
 
-	[RequiresUnreferencedCode(TrimmingWarningMessage)]
+
 	private static object ConvertValue(
-		[DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.All)]
-		Type type,
+
+			Type type,
 		string value, string path)
 	{
 		object result;
@@ -622,8 +646,7 @@ public static class ConfigurationBinder
 
 	private static Type FindOpenGenericInterface(
 		Type expected,
-		[DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.Interfaces)]
-		Type actual)
+			Type actual)
 	{
 		if (actual.IsGenericType &&
 			actual.GetGenericTypeDefinition() == expected)
@@ -640,12 +663,12 @@ public static class ConfigurationBinder
 				return interfaceType;
 			}
 		}
-		return null;
+		return null!;
 	}
 
 	private static List<PropertyInfo> GetAllProperties(
-		[DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.All)]
-		Type type)
+
+			Type type)
 	{
 		var allProperties = new List<PropertyInfo>();
 
@@ -659,7 +682,7 @@ public static class ConfigurationBinder
 		return allProperties;
 	}
 
-	[RequiresUnreferencedCode(PropertyTrimmingWarningMessage)]
+
 	private static object GetPropertyValue(PropertyInfo property, object instance, IConfiguration config, BinderOptions options)
 	{
 		string propertyName = GetPropertyName(property);
@@ -695,7 +718,7 @@ public static class ConfigurationBinder
 			string name = attributeData
 				.ConstructorArguments[0]
 				.Value?
-				.ToString();
+				.ToString()!;
 
 			return !string.IsNullOrWhiteSpace(name) ? name : property.Name;
 		}
@@ -703,3 +726,6 @@ public static class ConfigurationBinder
 		return property.Name;
 	}
 }
+
+
+
